Divided the unittest of the transition gait to the gait selection test and refactored the validate trajectory function in the transition_subgait.py to improve readability.
