+++ march_gait_selection/test/run_tests
@@ -2,10 +2,11 @@
 import rosunit
 
 from nosetests.test_gait_transition import TestTransitionTrajectory
-
+from nosetests.test_gait_selection import TestGaitSelection
 
 PKG = 'march_gait_selection'
 
 
 if __name__ == '__main__':
-    rosunit.unitrun(PKG, 'test_transition', TestTransitionTrajectory)
+    rosunit.unitrun(PKG, 'test_gait_transition', TestTransitionTrajectory)
+    rosunit.unitrun(PKG, 'test_gait_selection', TestGaitSelection)
